refactor(maestro): move import target tests into a sibling module
